Fix the config_file path for use as subdirectory
If another CMake project uses KinesisVideoProducerCpp by including
amazon-kinesis-video-streams-producer-sdk-cpp through an `add_subdirectory`
call, then using CMAKE_SOURCE_DIR points to an invalid path. Since `ROOT`
is set in the root CMakeLists.txt, we can use that variable to help point
to the correct path both when building the project on its own, as well
when the project is included in a parent CMake project.

This simple change allows a parent CMakeLists.txt to add this project and
build it successfully.